refactor(search): Move download logic into DownloadResults component
The full download flow (fetch, progress tracking, blob creation, trigger)
now lives in DownloadResults.tsx. SearchResults renders it with just
traceSummaries and rawTraces — no state or callbacks needed at the
parent level. createBlob and getStripCircular move with it.

pooledMap updated to use the iterator protocol (items.entries()) which
is more idiomatic and avoids any ambiguity about index assignment under
error paths.
