+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/createBlob.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-import { createBlob } from './index';
+import { createBlob } from './DownloadResults';
 import transformTraceData from '../../../model/transform-trace-data';
 import { TraceData, SpanData, SpanReference } from '../../../types/trace';
